third_party/WebKit/LayoutTests/http/tests/storage/async_cookies/document_get_set.html

<!doctype html>
<meta charset="utf-8">
<title>Async Cookes: cokieStore.getAll() sees cookieStore.set() cookie</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async testCase => {
  const tenYears = 10 * 365 * 24 * 60 * 60 * 1000;
  const expiresAt = Date.now() + tenYears;

  await cookieStore.set('cookie-name', 'cookie-value', { expires: expiresAt });
  const cookie = await cookieStore.get();

  assert_equals(cookies.length, 1);
  assert_equals(cookies[0].name, 'cookie-name');
  assert_equals(cookies[0].value, 'cookie-value');
}, 'cookieStore.get returns the cookie written by cookieStore.set');

</script>
